+++ ci4/app/Interfaces/QuickCommandList.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property QuickCommand[] $values
+ */
+interface QuickCommandList {
+
+}