+++ vue/src/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateQuickCommandDialog.vue
@@ -1,0 +1,119 @@
+<script setup lang="ts">
+import {computed, defineComponent, onMounted, onUnmounted, reactive, ref, watch} from 'vue'
+import type {DialogEventsInterface} from "@/components/Dialogs/DialogEventsInterface";
+
+export interface DeploymentSpecificationUpdateQuickCommandDialog_Input {
+    quickCommand: {
+        name: string;
+        command: string;
+    };
+
+    onSaveCallback: () => void;
+}
+
+const props = defineProps<{ input: DeploymentSpecificationUpdateQuickCommandDialog_Input, events: DialogEventsInterface }>();
+
+const used = ref(false);
+const showDialog = ref(false);
+const name = ref('');
+const command = ref('');
+
+// <editor-fold desc="Functions">
+
+onMounted(() => {
+    if (used.value) {
+        return;
+    }
+    used.value = true;
+    render();
+});
+
+onUnmounted(() => {
+});
+
+function render() {
+    name.value = props.input.quickCommand.name ?? '';
+    command.value = props.input.quickCommand.command ?? '';
+    showDialog.value = true;
+}
+
+function close() {
+    showDialog.value = false;
+    props.events.onClose();
+}
+
+// </editor-fold>
+
+// <editor-fold desc="View Binding Functions">
+
+function onSaveBtnClicked() {
+    props.input.quickCommand.name = name.value;
+    props.input.quickCommand.command = command.value;
+    props.input.onSaveCallback();
+    close();
+}
+
+function onCloseBtnClicked() {
+    close();
+}
+
+// </editor-fold>
+
+</script>
+
+<template>
+    <v-dialog
+        persistent
+        width="60vw"
+        v-model="showDialog">
+        <v-card
+            class="w-100 h-100">
+            <v-card-title>Quick Command</v-card-title>
+            <v-divider/>
+            <v-card-text>
+                <v-row>
+                    <v-col cols="12">
+                        <v-text-field
+                            v-model="name"
+                            variant="outlined"
+                            label="Name"
+                            hide-details
+                        />
+                    </v-col>
+                    <v-col cols="12">
+                        <v-text-field
+                            v-model="command"
+                            variant="outlined"
+                            label="Command"
+                            hide-details
+                        />
+                    </v-col>
+                </v-row>
+            </v-card-text>
+            <v-divider/>
+            <v-card-actions>
+                <v-spacer/>
+                <v-btn
+                    variant="tonal"
+                    color="grey"
+                    prepend-icon="fa fa-circle-xmark"
+                    @click="onCloseBtnClicked">
+                    Close
+                </v-btn>
+
+                <v-btn
+                    flat
+                    variant="tonal"
+                    prepend-icon="fa fa-check"
+                    color="green"
+                    @click="onSaveBtnClicked">
+                    Done
+                </v-btn>
+            </v-card-actions>
+        </v-card>
+    </v-dialog>
+</template>
+
+<style scoped>
+
+</style>