+++ vue/src/core/services/Deploy/models/DeploymentSpecificationQuickCommand.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 07-01-2024.
+ * Time: 08:28.
+ */
+import {DeploymentSpecificationQuickCommandDefinition} from './definitions/DeploymentSpecificationQuickCommandDefinition';
+
+export class DeploymentSpecificationQuickCommand extends DeploymentSpecificationQuickCommandDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}