+++ vue/src/core/services/Deploy/models/definitions/DeploymentSpecificationQuickCommandDefinition.ts
@@ -1,0 +1,77 @@
+/**
+ * Created by ModelParser
+ */
+import {DeploymentSpecification} from '../DeploymentSpecification';
+import {User} from '../User';
+import {BaseModel} from '../BaseModel';
+
+export class DeploymentSpecificationQuickCommandDefinition extends BaseModel {
+    deployment_specification_id?: number;
+    deployment_specification?: DeploymentSpecification;
+    name?: string;
+    command?: string;
+    id?: number;
+    created?: string;
+    updated?: string;
+    created_by_id?: number;
+    created_by?: User;
+    updated_by_id?: number;
+    updated_by?: User;
+
+    constructor(data?: any) {
+        super();
+        this.populate(data);
+    }
+
+    public populate(data?: any, patch = false) {
+        if (!patch) {
+            delete this.deployment_specification_id;
+            delete this.deployment_specification;
+            delete this.name;
+            delete this.command;
+            delete this.id;
+            delete this.created;
+            delete this.updated;
+            delete this.created_by_id;
+            delete this.created_by;
+            delete this.updated_by_id;
+            delete this.updated_by;
+        }
+
+        if (!data) return;
+        if (data.deployment_specification_id != null) {
+            this.deployment_specification_id = data.deployment_specification_id;
+        }
+        if (data.deployment_specification != null) {
+            this.deployment_specification = new DeploymentSpecification(data.deployment_specification);
+        }
+        if (data.name != null) {
+            this.name = data.name;
+        }
+        if (data.command != null) {
+            this.command = data.command;
+        }
+        if (data.id != null) {
+            this.id = data.id;
+        }
+        if (data.created != null) {
+            this.created = data.created;
+        }
+        if (data.updated != null) {
+            this.updated = data.updated;
+        }
+        if (data.created_by_id != null) {
+            this.created_by_id = data.created_by_id;
+        }
+        if (data.created_by != null) {
+            this.created_by = new User(data.created_by);
+        }
+        if (data.updated_by_id != null) {
+            this.updated_by_id = data.updated_by_id;
+        }
+        if (data.updated_by != null) {
+            this.updated_by = new User(data.updated_by);
+        }
+    }
+
+}